pyproject.toml
[tool.poetry]
name = "rag-avec-kg"
version = "0.1.0"
description = ""
authors = ["Robin Guiavarch <robin.guiavarch@gmail.com>"]
readme = "README.md"
packages = [
  { include = "src" }
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
sentence-transformers = "2.7.0"
faiss-cpu = ">=1.11.0,<2.0.0"
langchain = ">=0.3.25,<0.4.0"
neo4j = ">=5.28.1,<6.0.0"
pyyaml = ">=6.0.2,<7.0.0"
pdfplumber = ">=0.11.6,<0.12.0"
pymupdf = ">=1.26.0,<2.0.0"
tqdm = ">=4.67.1,<5.0.0"
rich = ">=14.0.0,<15.0.0"
pandas = ">=2.2.3,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
nltk = ">=3.9.1,<4.0.0"
py2neo = ">=2021.2.4,<2022.0.0"
matplotlib = ">=3.10.3,<4.0.0"
whoosh = "^2.7.4"
streamlit = "^1.45.1"
openai = "^1.88.0"
gradio = "^5.34.0"
torch = "2.2.2"
transformers = "4.39.3"
numpy = "1.26.4"
seaborn = "^0.13.2"
levenshtein = "^0.27.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
